Dynamic Address Bar!
• The address bar (and title) now update when navigating to difference
scenes/projects on the site. This updates the browser history as well,
so the back and forward buttons will traverse/reverse through the site
as expected. This means the site can work with autofill features and
would appear to be subdirectories for all projects.
• Projects can be shortcut to since they appear to be in their own
subdirectory.
• Fixed image URL references (in both CSS and JS) using relative paths
instead of absolute paths. This is fixed and makes updates easier going
forward.
• Protocol and address are no longer hardcoded so features like the
redirect can be tested on a localhost without being redirected to the
live site. This will make implementing future features easier.